[1/3] base: make overlay permission check configurable
google added security check when changing app permissions
to check for apps running with enabled draw over apps
(SYSTEM_ALERT_WINDOW) permission.

Since OmniSwitch is such an app it will always trigger this
and requires users to remove permissions before changing
app permissions. After that the permission must be added back.

Since this is a PITA add a config that will disable the check
This is reducing security against tapjacking so make it
clear to the user what it means to disable it

TODO: maybe add a even more prominent alert when
enabling the value

Change-Id: I7f1c047709a990e7b6f2aaafc9a25586f822ea92
Signed-off-by: mydongistiny <jaysonedson@gmail.com>

Conflicts:
	core/java/android/provider/Settings.java